reduce :pedantic? to :warn if :abort is specified.
Reason: when checking if a dependency would suit an existing project
(e.g. my project has a map of data somewhere and I want to serialize
that data to JSON but want to try out different library) dependency
conflicts will prevent the REPL from starting if `pedantic?` is set to
`:abort`.

The user should be able to add the dependency, try it out and handle any
such conflicts at a later point in time which is why this commit will
reduce the `:pedantic?` level to `:warn` if it is `:abort`.